basic headline parsing works
I've settled for specifying the generations manually. This is more
cumbersome to update than "smarter" implementations. However, the
logic is incredibly simple and it's probably faster than the "smarter"
implementations because it doesn't require state.

It's probably a good idea at some point to write code to generate
these rules, especially since we will eventually need to support past
10 headlines. This is another problem with this implementation: it
will never support arbitrary headline nesting. However, we could
probably just set it to a maximum of 100 and then adjust if anyone
feels this is limiting.